worker/pyproject.toml
[project]
name = "tellequant-worker"
version = "0.1.0"
description = "Tellequant voice agent worker — Pipecat + Twilio Media Streams + Deepgram + Gemini"
requires-python = ">=3.11"
dependencies = [
  "pipecat-ai[daily,deepgram,cartesia,google,openai,silero]>=0.0.47",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "twilio>=9.0",
  "loguru>=0.7",
  "websockets>=13.0",
]

[project.scripts]
tellequant-worker = "tellequant_worker.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tellequant_worker*"]
